+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2026-03-30
+
+### Added
+- Shadow Color and Shadow Opacity controls to BoxShadow
+- Shadow Color and Shadow Opacity controls to DropShadow
+- Updated `globalEffects` hook to apply box shadow color and opacity classes
+- Updated `globalFilters` hook to apply drop shadow color and opacity classes
+
 ## [1.2.1] - 2026-01-09
 
 ### Changed